+++ python/dalex/dalex/_explainer/object.py
@@ -191,7 +191,7 @@ class Explainer:
                         grid_points=101,
                         variable_splits=None,
                         processes=1,
-                        disable=False):
+                        verbose=False):
 
         """Creates CeterisParibus object
 
@@ -202,7 +202,7 @@ class Explainer:
         :param grid_points: number of points in a single variable split if calculated automatically
         :param variable_splits: mapping of variables into points the profile will be calculated, if None then calculate with the function `_calculate_variable_splits`
         :param processes: integer, number of parallel processes, iterated over variables
-        :param disable: disable tqdm progress bar printing
+        :param verbose: disable tqdm progress bar printing
         :return CeterisParibus object
         """
 
@@ -217,7 +217,7 @@ class Explainer:
                 processes=processes
             )
 
-        predict_profile_.fit(self, new_observation, y, disable)
+        predict_profile_.fit(self, new_observation, y, verbose)
 
         return predict_profile_
 
@@ -302,7 +302,7 @@ class Explainer:
                       intercept=True,
                       processes=1,
                       random_state=None,
-                      disable=False):
+                      verbose=False):
 
         """Dataset Level Variable Effect as Partial Dependency Profile or Accumulated Local Effects
 
@@ -317,7 +317,7 @@ class Explainer:
         :param intercept: False if center data on 0
         :param processes: integer, number of parallel processes, iterated over variables
         :param random_state: int, seed for random number generator
-        :param disable: disable tqdm progress bar printing
+        :param verbose: disable tqdm progress bar printing
         :return: VariableEffect object
         """
 
@@ -335,7 +335,7 @@ class Explainer:
         I = np.random.choice(np.arange(N), N, replace=False)
 
         ceteris_paribus = CeterisParibus(grid_points=grid_points, processes=processes)
-        ceteris_paribus.fit(self, self.data.iloc[I, :], self.y[I], disable=disable)
+        ceteris_paribus.fit(self, self.data.iloc[I, :], self.y[I], verbose=verbose)
 
         model_profile_ = AggregatedProfiles(
             type=type,
@@ -347,7 +347,7 @@ class Explainer:
             random_state=random_state
         )
 
-        model_profile_.fit(ceteris_paribus, disable)
+        model_profile_.fit(ceteris_paribus, verbose)
 
         return model_profile_