+++ python/dalex/dalex/dataset_level/_aggregated_profiles/object.py
@@ -50,7 +50,7 @@ class AggregatedProfiles:
 
     def fit(self,
             ceteris_paribus,
-            disable=False):
+            verbose=False):
 
         # are there any other cp?
         from dalex.instance_level import CeterisParibus
@@ -80,7 +80,7 @@ class AggregatedProfiles:
         all_profiles = create_x(all_profiles, self.variable_type)
 
         self.result = aggregate_profiles(all_profiles, self.type, self.groups, self.intercept,
-                                         self.span, disable)
+                                         self.span, verbose)
 
         self.mean_prediction = all_observations['_yhat_'].mean()