+++ python/dalex/dalex/instance_level/_ceteris_paribus/utils.py
@@ -10,7 +10,7 @@ def calculate_ceteris_paribus(explainer,
                               variable_splits,
                               y,
                               processes,
-                              disable=False):
+                              verbose=False):
     """
     Inner method that calculates ceteris paribus and some additional fields.
 
@@ -24,7 +24,7 @@ def calculate_ceteris_paribus(explainer,
                                           new_observation,
                                           variable_splits,
                                           processes,
-                                          disable)
+                                          verbose)
 
     profiles.loc[:, '_label_'] = explainer.label
 
@@ -48,7 +48,7 @@ def calculate_variable_profile(predict_function,
                                data,
                                variable_splits,
                                processes,
-                               disable=False):
+                               verbose=False):
     """
     Inner function that calculates ceteris paribus for all variables.
 
@@ -60,7 +60,7 @@ def calculate_variable_profile(predict_function,
 
     if processes == 1:
         profile = []
-        for variable in tqdm(variable_splits, desc="Calculating ceteris paribus", disable=disable):
+        for variable in tqdm(variable_splits, desc="Calculating ceteris paribus", disable=verbose):
             split_points = variable_splits[variable]
             profile.append(single_variable_profile(predict_function, model, data, variable, split_points))
     else: