+++ python/dalex/NEWS.md
@@ -1,7 +1,7 @@
 dalex (development)
 ----------------------------------------------------------------
 * *fix:* `vertical_spacing` in `VariableImportance.plot` when `split='variable'`
-* *feature:* added `disable` parameter where `tqdm` is used to verbose progress bar
+* *feature:* added `verbose` parameter where `tqdm` is used to verbose progress bar
 * *feature:* added `loss_one_minus_auc` function that can be used with `loss_function='1-auc'` in `model_parts`
 * *defaults:* use different `loss_function` for classification and regression ([#248](https://github.com/ModelOriented/DALEX/issues/248))
 * *defaults:* models that use `proba` yhats now get `model_type='classification'` if it's not specified